Table to store last "topk update" timestamp of the query.
Separate table is required for this, as conditional update is required
to "lock" the update, non-counter field must be part of primary key in
Cassandra and conditional updates are not allowed for primary key fields.

We also add one extra counter to count queries since last update to allow
flush both on time and query count thresholds.